refactor(recipes): moved recipe filter building to separate functions
